All files modified in order to tackle some serious debugging... comments folder has been moved out of database into the intro_project folder and all tests are passing!!!
