+++ backend/src/main/java/our_search/intro_project/database/services/CommentService.java
@@ -4,13 +4,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import our_search.intro_project.database.entities.Comment;
 import our_search.intro_project.database.repositories.CommentRepository;
-import our_search.intro_project.comments.dto.FetchedComment;
-import our_search.intro_project.database.comments.dto.GetCommentRequest;
 
 
 import java.util.List;
 import java.util.Optional;
-import java.util.stream.Collectors;
 
 @Service
 public class CommentService {
@@ -51,14 +48,7 @@ public class CommentService {
         return false;
     }
 
-    public List<FetchedComment> getCommentsByPostId(Integer postId) {
-        return commentRepository.findByPostId(postId).stream()
-                .map(comment -> new FetchedComment(
-                        comment.getCommentId(),
-                        comment.getContents(),
-                        comment.getUser().getName(),
-                        comment.getCreatedAt()
-                ))
-                .collect(Collectors.toList());
+    public List<Comment> getCommentsByPostId(Integer postId) {
+        return commentRepository.findByPostId(postId);
     }
 }
